Script to compare numerical solvers.
The script compares the performance of the numerical solvers
on the first-order system described by the differential
equation x' = -x.

This is simple enough to compute the analytic solution such that
it can be used to assess the accuracy of the numerical methods,
in our case `Euler`, `Improved-Euler` and `Runge-Kutta`.